pyproject.toml
[tool.poetry]
name = "fastapi-starter"
version = "0.1.0"
description = ""
authors = ["daitasu"]

[tool.poetry.dependencies]
python="3.10.2"
fastapi = "^0.74.0"
uvicorn = "^0.17.5"
SQLAlchemy = "^1.4.31"
psycopg2 = "^2.9.3"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
